Make hard test timeouts more gentle
When we're hitting hard test timeouts we never have logs from the test
case that caused it because the whole test runner is killed by a
signal. Instead of a direct hard kill use a gentle timeout and give
the test case some time to exit cleanly until we hit a hard test
timeout. This should make debugging such things a lot easier.

Change-Id: Ia06bd66c24041a7e08639812a8fc4788bc2b9683