Fix test race in test_periodic_override
The test case test_periodic_override checks if project overrice in
timer triggered pipelines works. For this it enables the timer trigger
and then creates the branch stable/havana. After some time it checks
all in flight builds that they checked out the correct
revision. However the branch creation itself first creates a branch
and then a commit. If the timer kicks in between those two steps zuul
will use the same commit for both branches and the test fails.

Fix this by creating the branch before enabling the timer.

Change-Id: I21866978f4ffaaad4484b2dc93c6a6afa0745f9d