Use xterm.js for live log streaming
We currently use a very simplistic self written mechanism for live log
streaming. This has severe performance drawbacks when dealing with
large live logs. The component xterm.js widely used out there that is
specialized for this task and handles huge logs just fine.

We also don't need the autoscroll checkbox anymore as this is handled
automagically by xterm.js. It stops following the stream when
scrolling and starts following again after scrolling to the bottom.

Further it makes it easy to have clickable links in the live log.

Change-Id: I3492e983bf248b4f286edc1bf9db3d52297da993