Whitelist pageup and down keys in xterm
We currently block all keys in xterm so we can use ctrl+c for copying
text. However this blocks scrolling via shift+pageup and
down. Whitelist those buttons so scrolling per keyboard works.

This blocks scrolling per keyboard. In case this is still needed we
should add it back with a specific blacklist or whitelist.

Change-Id: Ida8c90f556912941af528d3c1fd322cf1256185d