Check branches in supercedent pipelines
Supercedent was designed for a post pipeline and predates common
use of a promote pipeline.  It therefore only consulted the 'ref'
attribute of a change (so that it would work on tags as well as
branches).  However, it should work just fine as a manager for
promote pipelines too, so have it check the branch attribute for
de-duplication first, and only use the ref attribute if it isn't
present.

Change-Id: I1393c828b86a577fba568a183efee98cf5d030e7