Fix dynamic loading of trusted layouts
There were special circumstances under which we would return a trusted
config layout as a valid dynamic layout update. Specifically if there
were existing layout errors in code unrelated to the trusted config
update we would fall through error handling such that the trusted config
would be returned.

Fix this by recording the trusted and untrusted layouts as separate
variables so that we can very clearly switch through our state cases and
be clear that we never return the trusted_layout. Instead of there is a
trusted_layout update that would otherwise be accepted we return the
current pipeline config.

Part of the fix here is to write out a "switch" table that clearly lists
and handles all known possible cases. If we hit an unknown case we
return an error. Hopefully by being extra clear about our intent we
avoid unexpected behavior like this in the future.

Story: 2005452
Change-Id: I095819bf2288b4101352badfaf0e0fa8062c2829